Refactored to use pycurl/libcurl if it's available
