+++ README.mdown
@@ -90,7 +90,6 @@ could use something like this:
 
     #!/usr/bin/env python
     
-    import asyncore
     import twitstream
     
     USER = 'test'
@@ -102,7 +101,7 @@ could use something like this:
     
     if __name__ == '__main__':
         # Call a specific API method from the twitstream module: 
-        twitstream.spritzer(USER, PASS, callback)
+        stream = twitstream.spritzer(USER, PASS, callback)
         
         # Loop forever on the streaming call:
-        asyncore.loop()
+        stream.run()
